engine: extract dispatch::abilities (#159)
Move activate_ability, pay_activation_costs, resolve_activated_ability,
and check_cost_payable verbatim out of dispatch/mod.rs into the new
dispatch/abilities.rs submodule. Back-patch reaction_windows.rs to call
super::abilities::resolve_activated_ability and
super::abilities::check_cost_payable. Remove orphaned imports from
mod.rs (card_registry, Cost, Trigger, CardInstanceId, Investigator,
apply_effect, EvalContext).
